Added context for when i did certain things
I also updated the CSS slightly to account for the addition of the C# language
A major revamp to the CSS will be done when I have some more free time. I'll also see about being able to remotely update the text whenever i deem necessary. (mainly the Status star text and the about me section. The Skills/Tools section would be a bit difficult to update thanks to the icons. same goes for the projects area.)
one of the major revamps will be switching grid with flexbox and dividing the header, footer and body css within their respective files.